+++ backend/src/controllers/emails.js
@@ -11,7 +11,6 @@ const table = {
 };
 
 export async function postEmail(req, res) {
-
     try {
         const { email } = req.body;
 
@@ -30,15 +29,16 @@ export async function postEmail(req, res) {
     } catch (error) {
         const routeError = new ErrorRoute(error, table).typeError();
         return res.status(routeError.code).json({ error: routeError.message, status: false });
-    }
+    };
 };
 
 export async function putEmail(req, res) {
     try {
         const id = Number(req.params.id) || 0;
         const { email } = req.body;
+
         if(id<=0){
-            return res.status(404).json({ message: `${table.table}. No encontrado`, status: false });
+            return res.status(401).json({ message: `${table.table}. Invalido`, status: false });
         };
 
         const {rowCount, rows} = await pool.query("SELECT * FROM emails WHERE id=$1", [
@@ -75,9 +75,15 @@ export async function putEmail(req, res) {
 export async function deleteEmail(req, res) {
     try {
         const id = Number(req.params.id);
+
+        if(id<=0){
+            return res.status(401).json({ message: `${table.table}. Invalido`, status: false });
+        };
+
         const { rowCount } = await pool.query("DELETE FROM emails where id = $1", [
             id,
         ]);
+        
         if (rowCount === 0){
             return res.status(404).json({ message: `${table.table}. No encontrado`, status: false });
         };
